requirements.txt
aiosqlite==0.22.1
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
attrs==25.4.0
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
cryptography==46.0.5
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.135.1
gitdb==4.0.12
GitPython==3.1.46
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
librt==0.8.1
MarkupSafe==3.0.3
mypy==1.19.1
mypy_extensions==1.1.0
narwhals==2.18.0
numpy==2.4.3
packaging==26.0
pandas==2.3.3
passlib==1.7.4
pathspec==1.0.4
pillow==12.1.1
pluggy==1.6.0
protobuf==6.33.5
pyarrow==23.0.1
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pytest==9.0.2
pytest-asyncio==1.3.0
python-dateutil==2.9.0.post0
python-jose==3.5.0
python-multipart==0.0.22
pytz==2026.1.post1
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
rsa==4.9.1
ruff==0.15.5
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
smmap==5.0.3
SQLAlchemy==2.0.47
starlette==0.52.1
streamlit==1.55.0
tenacity==9.1.4
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.5
types-passlib==1.7.7.20260211
types-pyasn1==0.6.0.20250914
types-python-jose==3.5.0.20250531
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.41.0
